deep merge
